Struct rquery::Document

pub struct Document {
    // some fields omitted
}

The DOM tree representation of the parsed document.

Methods

impl Document

fn new_from_xml_stream<R: Read>(stream: R) -> Result<Document, DocumentError>

Creates a new document from a byte stream.

fn new_from_xml_string(string: &str) -> Result<Document, DocumentError>

Creates a new document from a string.

fn new_from_xml_file(filename: &str) -> Result<Document, DocumentError>

Creates a new document from a file.

fn number_of_elements(&self) -> usize

Returns the total number of elements in the document.

fn select_all<'a>(&'a self, selector: &str) -> Result<Box<Iterator<Item=&'a Element> + 'a>, SelectError>

Searches the document for elements matching the given CSS selector.

fn select<'a>(&'a self, selector: &str) -> Result<&'a Element, SelectError>

Just like select_all but only returns the first match.